- added Facade; - removed interfaces; - added RxJava, Guava and Lombok dependencies;
